+++ TeamInventum/build.release.gradle
@@ -1,0 +1,8 @@
+dependencies {
+    compile project(':FtcRobotController')
+    compile (name: 'RobotCore-release', ext: 'aar')
+    compile (name: 'Hardware-release', ext: 'aar')
+    compile (name: 'FtcCommon-release', ext: 'aar')
+    compile (name:'Analytics-release', ext:'aar')
+    compile (name:'WirelessP2p-release', ext:'aar')
+}
