+++ settings.gradle
@@ -3,3 +3,4 @@ include ':TeamCode'
 include ':Team7153'
 include ':Team11383'
 include ':Team11750'
+include ':TeamInventum'
